Felix's edits to views.py

+++ views.py
@@ -370,21 +370,23 @@ def loadSynthesisFromId(request):
 @login_required
 def renderSynthesis(request):
     profile = request.user.profile
-    #Sometimes, the user doesn't even have a previous problem, so deleting doesn't always work.
-    try:
-        for arrowModel in profile.currentSynthesisProblem.arrows.all():
-            arrowModel.delete()
-        for moleculeModel in profile.currentSynthesisProblem.molecules.all():
-            moleculeModel.delete()
-        for arrowModel in profile.currentSynthesisProblem.solution.arrows.all():
-            arrowModel.delete()
-        for moleculeModel in profile.currentSynthesisProblem.solution.molecules.all():
-            moleculeModel.delete()
-        profile.currentSynthesisProblem.solution.delete()
-        profile.currentSynthesisProblem.target.delete()
-        profile.currentSynthesisProblem.delete()
-    except:
-        pass
+    #If the retain attribute is True, don't delete.
+    if not(profile.currentSynthesisProblem.retain):
+        #Sometimes, the user doesn't even have a previous problem, so deleting doesn't always work.
+        try:
+            for arrowModel in profile.currentSynthesisProblem.arrows.all():
+                arrowModel.delete()
+            for moleculeModel in profile.currentSynthesisProblem.molecules.all():
+                moleculeModel.delete()
+            for arrowModel in profile.currentSynthesisProblem.solution.arrows.all():
+                arrowModel.delete()
+            for moleculeModel in profile.currentSynthesisProblem.solution.molecules.all():
+                moleculeModel.delete()
+            profile.currentSynthesisProblem.solution.delete()
+            profile.currentSynthesisProblem.target.delete()
+            profile.currentSynthesisProblem.delete()
+        except:
+            pass
     
     modes = checkboxUpdate(request)
     if modes == []:
